* commit '1e82d6e101e18437e50a603ad958d0543b2134d4':
  bundle zlib on windows
  Check for -mrelax-relocations=no support
  Test static-libs-image target
  Remove zlib bundling
  Fix incorrect renames
  Discard version number changes
  Remove our static lib changes
  Discard unnecessary edits
  Sync with jdk11-graalvm tip after 8231985

+++ src/hotspot/share/compiler/disassembler.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2008, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2008, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -72,7 +72,7 @@ bool Disassembler::load_library() {
   char* p = strrchr(buf, '/');
   *p = '\0';
   strcat(p, "/lib/");
-  lib_offset = jvm_offset = (int)strlen(buf);
+  lib_offset = jvm_offset = strlen(buf);
 #else
   {
     // Match "jvm[^/]*" in jvm_path.
